requirements.txt
aiohttp==3.8.3
pyserial==3.5
gunicorn==20.1.0
